service: need postfix for configuration to install as private service only
The OE webos_system_class bbclass determines the install location for the service
configuration based on it's postfix. If there is no prv/pub postfix the service
configuration installed for both public and private which isn't what we want for the
update service which should be only available on the private bus.

Signed-off-by: Simon Busch <morphis@gravedo.de>